requirements.txt
streamlit
pandas
snowflake-connector-python
